+++ internal/forge/gitea/api_connector.go
@@ -1,0 +1,28 @@
+package gitea
+
+import (
+	"code.gitea.io/sdk/gitea"
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+)
+
+func (self Connector) VerifyConnection() forgedomain.VerifyConnectionResult {
+	user, _, err := self.client.GetMyUserInfo()
+	if err != nil {
+		return forgedomain.VerifyConnectionResult{
+			AuthenticatedUser:   None[string](),
+			AuthenticationError: err,
+			AuthorizationError:  nil,
+		}
+	}
+	_, _, err = self.client.ListRepoPullRequests(self.Organization, self.Repository, gitea.ListPullRequestsOptions{
+		ListOptions: gitea.ListOptions{
+			PageSize: 1,
+		},
+	})
+	return forgedomain.VerifyConnectionResult{
+		AuthenticatedUser:   NewOption(user.UserName),
+		AuthenticationError: nil,
+		AuthorizationError:  err,
+	}
+}
